refactor: add generic file iteration trait
